WIP PyIceberg support
