+++ adapta/storage/distributed_object_store/v2/datastax_astra/astra_client.py
@@ -67,7 +67,8 @@ from cassandra.query import dict_factory, BatchType  # pylint: disable=E0611
 
 from adapta import __version__
 from adapta.storage.distributed_object_store.v2.datastax_astra._models import SimilarityFunction, VectorSearchQuery
-from adapta.storage.models.filter_expression import Expression, AstraFilterExpression, compile_expression
+from adapta.storage.models.expression_dsl.filter_expression import Expression, compile_expression
+from adapta.storage.models.expression_dsl.astra_filter_expression import AstraFilterExpression
 from adapta.utils import chunk_list, rate_limit
 
 TModel = TypeVar("TModel")  # pylint: disable=C0103
