+++ adapta/storage/distributed_object_store/v3/datastax_astra/_models.py
@@ -20,12 +20,12 @@ import re
 from enum import Enum
 from typing import Any
 
-from adapta.storage.models.filter_expression import (
+from adapta.storage.models.expression_dsl.filter_expression import (
     Expression,
     compile_expression,
-    AstraFilterExpression,
     FilterExpressionOperation,
 )
+from adapta.storage.models.expression_dsl.astra_filter_expression import AstraFilterExpression
 
 
 class SimilarityFunction(Enum):
