+++ adapta/storage/models/expression_dsl/filter_expression.py
@@ -1,15 +1,15 @@
 """
     Models for generating filter expressions for PyArrow and Astra.
 """
-import datetime
-import math
 from abc import ABC, abstractmethod
 from typing import final, Any, TypeVar, Generic, Self
 from enum import Enum
 import pyarrow.compute
-from pyarrow.dataset import field as pyarrow_field
 
-from adapta.utils import chunk_list
+try:
+    import pyiceberg.expressions
+except (ImportError, ModuleNotFoundError):
+    pass
 
 TCompileResult = TypeVar("TCompileResult")  # pylint: disable=invalid-name
 
@@ -25,14 +25,46 @@ class FilterExpressionOperation(Enum):
         "astra": lambda left_exprs, right_exprs: [
             left_expr | right_expr for left_expr in left_exprs for right_expr in right_exprs
         ],
+        "trino": "AND",
+        "iceberg": pyiceberg.expressions.And,
     }
-    OR = {"arrow": pyarrow.compute.Expression.__or__, "astra": lambda left_exprs, right_exprs: left_exprs + right_exprs}
-    GT = {"arrow": pyarrow.compute.Expression.__gt__, "astra": "__gt"}
-    GE = {"arrow": pyarrow.compute.Expression.__ge__, "astra": "__gte"}
-    LT = {"arrow": pyarrow.compute.Expression.__lt__, "astra": "__lt"}
-    LE = {"arrow": pyarrow.compute.Expression.__le__, "astra": "__lte"}
-    EQ = {"arrow": pyarrow.compute.Expression.__eq__, "astra": ""}
-    IN = {"arrow": pyarrow.compute.Expression.isin, "astra": "__in"}
+    OR = {
+        "arrow": pyarrow.compute.Expression.__or__,
+        "astra": lambda left_exprs, right_exprs: left_exprs + right_exprs,
+        "trino": "OR",
+        "iceberg": pyiceberg.expressions.Or,
+    }
+    GT = {
+        "arrow": pyarrow.compute.Expression.__gt__,
+        "astra": "__gt",
+        "trino": ">",
+        "iceberg": pyiceberg.expressions.GreaterThan,
+    }
+    GE = {
+        "arrow": pyarrow.compute.Expression.__ge__,
+        "astra": "__gte",
+        "trino": ">=",
+        "iceberg": pyiceberg.expressions.GreaterThanOrEqual,
+    }
+    LT = {
+        "arrow": pyarrow.compute.Expression.__lt__,
+        "astra": "__lt",
+        "trino": "<",
+        "iceberg": pyiceberg.expressions.LessThan,
+    }
+    LE = {
+        "arrow": pyarrow.compute.Expression.__le__,
+        "astra": "__lte",
+        "trino": "<=",
+        "iceberg": pyiceberg.expressions.LessThanOrEqual,
+    }
+    EQ = {
+        "arrow": pyarrow.compute.Expression.__eq__,
+        "astra": "",
+        "trino": "=",
+        "iceberg": pyiceberg.expressions.EqualTo,
+    }
+    IN = {"arrow": pyarrow.compute.Expression.isin, "astra": "__in", "trino": "IN", "iceberg": pyiceberg.expressions.In}
 
     def to_string(self):
         """
@@ -264,175 +296,6 @@ class FilterExpression(Generic[TCompileResult], ABC):
 
         # Combine the compiled left and right operands
         return self._combine_results(compiled_left, compiled_right, expr.operation)
-
-
-@final
-class AstraFilterExpression(FilterExpression[list[dict[str, Any]]]):
-    """
-    A concrete implementation of the 'FilterExpression' abstract class for Astra.
-    """
-
-    # This value represents the threshold for the maximum length of a list in an IN filter in Astra
-    in_select_cartesian_product_failure_threshold = 25
-
-    def _compile_base_case(
-        self, field_name: str, field_values: Any, operation: FilterExpressionOperation
-    ) -> TCompileResult:
-        if (
-            operation == FilterExpressionOperation.IN
-            and isinstance(field_values, list)
-            and len(field_values) > self.in_select_cartesian_product_failure_threshold
-        ):
-            return self._isin_large_list_result(field_name, field_values, operation)
-        return [{f"{field_name}{operation.value['astra']}": field_values}]
-
-    def _combine_results(
-        self, compiled_result_a: TCompileResult, compiled_result_b: TCompileResult, operation: FilterExpressionOperation
-    ) -> TCompileResult:
-        return operation.value["astra"](compiled_result_a, compiled_result_b)
-
-    def _isin_large_list_result(
-        self, field_name: str, field_values: list[Any], operation: FilterExpressionOperation
-    ) -> TCompileResult:
-        # Compile each chunk into an IN operation expression
-        return [
-            {f"{field_name}{operation.value['astra']}": chunk}
-            for chunk in chunk_list(
-                field_values, math.ceil(len(field_values) / self.in_select_cartesian_product_failure_threshold)
-            )
-        ]
-
-
-@final
-class ArrowFilterExpression(FilterExpression[pyarrow.compute.Expression]):
-    """
-    A concrete implementation of the 'FilterExpression' abstract class for PyArrow.
-    """
-
-    def _compile_base_case(
-        self, field_name: str, field_values: Any, filter_operation: FilterExpressionOperation
-    ) -> TCompileResult:
-        field, field_values = self._handle_nested_types(
-            pyarrow_field(field_name),
-            field_values=field_values,
-            filter_operation=filter_operation,
-        )
-
-        return filter_operation.value["arrow"](field, field_values)
-
-    @staticmethod
-    def _handle_nested_types(
-        field: pyarrow.compute.Expression,
-        field_values: Any,
-        filter_operation: FilterExpressionOperation,
-        separator: str = ",",
-    ) -> tuple[pyarrow.compute.Expression, Any]:
-        """
-        Handle nested types in PyArrow filter expressions.
-
-        Processes complex nested types, such as lists and dictionaries, by converting
-        the values to strings and joining them with a specified separator. If the
-        field values are not nested types, they are returned unchanged.
-
-        :param field: The PyArrow field expression to process.
-        :param field_values: The values to compare, which may be nested types
-            (e.g., lists, dictionaries) or plain values.
-        :param filter_operation: The filtering operation to apply.
-        :param separator: The separator used for joining nested values. Defaults to ",".
-        :returns: A tuple containing the processed field expression and transformed
-            field values.
-        """
-
-        def join_list_of_lists(values: list[list], sep: str) -> list[str]:
-            return [sep.join(map(str, inner_list)) for inner_list in values]
-
-        def join_dict_values(values: list[dict] | dict, sep: str) -> str | list[str]:
-            if isinstance(values, dict):
-                return sep.join(map(str, values.values()))
-            return [sep.join(map(str, d.values())) for d in values]
-
-        # Handle list of lists
-        if isinstance(field_values, list) and isinstance(field_values[0], list):
-            field_values = join_list_of_lists(field_values, separator)
-            field = pyarrow.compute.binary_join(
-                field.cast(pyarrow.list_(pyarrow.string())), pyarrow.scalar(separator, pyarrow.large_string())
-            )
-
-        # Handle list of dicts
-        elif isinstance(field_values, list) and isinstance(field_values[0], dict):
-            field = pyarrow.compute.binary_join_element_wise(
-                *[
-                    pyarrow.compute.struct_field(field, key).cast(pyarrow.large_string())
-                    for key in field_values[0].keys()
-                ],
-                pyarrow.scalar(separator, pyarrow.large_string()),
-            )
-            field_values = join_dict_values(field_values, separator)
-
-        # Handle list comparison with EQUALS
-        elif isinstance(field_values, list) and filter_operation == FilterExpressionOperation.EQ:
-            field_values = separator.join(map(str, field_values))
-            field = pyarrow.compute.binary_join(
-                field.cast(pyarrow.list_(pyarrow.string())), pyarrow.scalar(separator, pyarrow.large_string())
-            )
-
-        # Handle dict
-        elif isinstance(field_values, dict):
-            field = pyarrow.compute.binary_join_element_wise(
-                *[pyarrow.compute.struct_field(field, key).cast(pyarrow.large_string()) for key in field_values.keys()],
-                pyarrow.scalar(separator, pyarrow.large_string()),
-            )
-            field_values = join_dict_values(field_values, separator)
-
-        return field, field_values
-
-    def _combine_results(
-        self,
-        compiled_result_a: TCompileResult,
-        compiled_result_b: TCompileResult,
-        filter_operation: FilterExpressionOperation,
-    ) -> TCompileResult:
-        return filter_operation.value["arrow"](compiled_result_a, compiled_result_b)
-
-
-class TrinoFilterExpression(FilterExpression[str]):
-    """
-    A concrete implementation of the 'FilterExpression' abstract class for Trino SQL.
-    Compiles filter expressions into Trino-compatible SQL WHERE clause fragments.
-    """
-
-    def _compile_base_case(self, field_name: str, field_values: Any, operation: FilterExpressionOperation) -> str:
-        # Map EQ to '=' for Trino
-        if operation == FilterExpressionOperation.EQ:
-            return f"{field_name} = {self._format_value(field_values)}"
-
-        # Handle IN as a series of ORs for Trino
-        if operation == FilterExpressionOperation.IN:
-            if not isinstance(field_values, list):
-                raise ValueError("IN operation requires a list of values")
-            return f"{field_name} IN ({', '.join(self._format_value(v) for v in field_values)})"
-        # Handle other operations
-        op_str = operation.to_string()
-        return f"{field_name} {op_str} {self._format_value(field_values)}"
-
-    def _combine_results(
-        self, compiled_result_a: str, compiled_result_b: str, operation: FilterExpressionOperation
-    ) -> str:
-        op_str = operation.to_string()
-        return f"({compiled_result_a} {op_str} {compiled_result_b})"
-
-    @staticmethod
-    def _format_value(value: Any) -> str:
-        # Format value for SQL: quote strings, leave numbers as is and return NULL for None
-        if isinstance(value, str):
-            return f"'{value}'"
-        if isinstance(value, datetime.datetime):
-            return f"TIMESTAMP '{value}'"
-        if isinstance(value, datetime.date):
-            return f"DATE '{value}'"
-        if value is None:
-            return "NULL"
-        return str(value)
 
 
 def compile_expression(expression: Expression, target: type[FilterExpression[TCompileResult]]) -> TCompileResult:
